Fixed bugs

+++ FarmNer/Assets/Flashlight/Flashlight.gd
@@ -1,7 +1,7 @@
 extends Node2D
 
 var mousePos
-var flashlightOn = true
+var flashlightOn = false
 
 func _process(delta):
 	mousePos = get_local_mouse_position()
@@ -23,29 +23,31 @@ func turnOnOffFlashlight():
 
 # Focus' flashlight's beam and plays and stops sound
 func attackBeam():
-	if (Input.is_action_pressed("Attack")):
-		if not $FlashLightBeamSound/AudioStreamPlayer2D.playing:
-			$FlashLightBeamSound/AudioStreamPlayer2D.volume_db = -15.0
-			$FlashLightBeamSound/AudioStreamPlayer2D.play()
+	rotation += mousePos.angle() * 0.5
+	if (flashlightOn == false):
+		
+		if (Input.is_action_pressed("Attack")):
+			if not $FlashLightBeamSound/AudioStreamPlayer2D.playing:
+				$FlashLightBeamSound/AudioStreamPlayer2D.volume_db = -15.0
+				$FlashLightBeamSound/AudioStreamPlayer2D.play()
 			
-		$Attack.monitorable = true
+			$Attack.monitorable = true
 		
-		$Tween.remove_all()
-		$Tween.interpolate_property($Sprite/Beam, "scale", $Sprite/Beam.scale, Vector2(4.941, 3.0), 1.0, Tween.TRANS_LINEAR, Tween.EASE_OUT_IN)
-		$Tween.interpolate_property($Sprite/Beam, "energy", $Sprite/Beam.energy, 2.5, 0.5, Tween.TRANS_LINEAR, Tween.EASE_OUT_IN)
-		$Tween.start()
-		rotation += mousePos.angle() * 0.09
-	else:
-		if $FlashLightBeamSound/AudioStreamPlayer2D.playing == true:
-			$Tween.interpolate_property($FlashLightBeamSound/AudioStreamPlayer2D, "volume_db", $FlashLightBeamSound/AudioStreamPlayer2D.volume_db, -80.0, 0.2, Tween.TRANS_LINEAR, Tween.EASE_IN_OUT)
+			$Tween.remove_all()
+			$Tween.interpolate_property($Sprite/Beam, "scale", $Sprite/Beam.scale, Vector2(4.941, 3.0), 1.0, Tween.TRANS_LINEAR, Tween.EASE_OUT_IN)
+			$Tween.interpolate_property($Sprite/Beam, "energy", $Sprite/Beam.energy, 2.5, 0.5, Tween.TRANS_LINEAR, Tween.EASE_OUT_IN)
 			$Tween.start()
+			rotation += mousePos.angle() * 0.09
+		else:
+			if $FlashLightBeamSound/AudioStreamPlayer2D.playing == true:
+				$Tween.interpolate_property($FlashLightBeamSound/AudioStreamPlayer2D, "volume_db", $FlashLightBeamSound/AudioStreamPlayer2D.volume_db, -80.0, 0.2, Tween.TRANS_LINEAR, Tween.EASE_IN_OUT)
+				$Tween.start()
 	
-		$Attack.monitorable = false
+			$Attack.monitorable = false
 		
-		$Tween.stop($Sprite/Beam)
-		$Tween.interpolate_property($Sprite/Beam, "scale", $Sprite/Beam.scale, Vector2(4.941, 7.682), 1.0, Tween.TRANS_LINEAR, Tween.EASE_OUT_IN)
-		$Tween.interpolate_property($Sprite/Beam, "energy", $Sprite/Beam.energy, 1.5, 0.5, Tween.TRANS_LINEAR, Tween.EASE_OUT_IN)
-		rotation += mousePos.angle() * 0.5
+			$Tween.stop($Sprite/Beam)
+			$Tween.interpolate_property($Sprite/Beam, "scale", $Sprite/Beam.scale, Vector2(4.941, 7.682), 1.0, Tween.TRANS_LINEAR, Tween.EASE_OUT_IN)
+			$Tween.interpolate_property($Sprite/Beam, "energy", $Sprite/Beam.energy, 1.5, 0.5, Tween.TRANS_LINEAR, Tween.EASE_OUT_IN)
 
 
 func _on_Tween_tween_completed(object, key):
